Shortened overall code, effectively removed saving
First of all, by removing a lot of the unnecessary abstractions that I wrote when I created this, the code is far shorter. Secondly, I believe that you shouldn't really be saving settings. A person will likely use different settings somewhat often, and it's only a minor inconvenience to not be able to save them. You can still save your keys though, albeit to a different file named "key.txt".

Personally, I don't the code is that bad anymore. There was a lot of repetition before, but less now. Seeing as the process is (whatever way you do it) highly procedural, this code is alright.